fix: Decouple Emacs text scale from current state
Make the configuration more deterministic by setting the scaling in a
manner that is not sensitive to the current scaling setting. Since there
is no scale setter where we can set a fixed scaling factor, we have to
reset and then increment to guarantee that every execution of the hook
will produce the same result.