fix: Disable Emacs default-text-scale-increment in new frames
The default-text-scale-increment configuration causes problems when
new frames are opened which some packages like Magit do e.g.: Magit
creates a new frame in order to gather the COMMIT_EDITMSG which
appears with text scaled needlessly large due to the additional
scale-increment operation previously in our make-frame-hook.

Disabling the increment allows us to to cirumvent the issue. Hunch is
that, in the Magit case (my problem), default-text-scale-reset is
invoked with a prefix which, instead of just resetting the scale,
actually sets the default scale and then executed the subsequent
increment operation which results to and increment on top of a
potentially already incremented scale.

Keeping default-text-scale-reset in the hook as is currently the case
is still problematic since every new frame triggers a global reset of
the default-text-scale (which will redraw all open frames). It may be
better after all to resort to using prefixed default-text-scale-reset
to set the default-text-scale for all subsequent frames after scaling
the text manually the first time I open the Emacs client.