chore: Config Emacs org-babel lang list in package chapter
Just to obtain some overview, it makes sense to manage Org Babel's capabilities
wherever the related packages themselves are loaded and configured. As a first
stage, I therefore disable the Org block in order to opt for a slightly
different design.

Now I should be able to navigate to a chapter and find most of the needed
settings right there -- now the org-babel stuff is part of that scope.