fix: decouple async org-roam sync from volatile org-roam-directory
Variable org-roam-directory changes quite a bit but we always want to run the
sync command from the root of our org-roam tree. As a result of this, we now set
org-roam-directory in the async process to the root Org-Roam directory. This
decouples it from variable org-roam-directory that is rather volatile as it can
be something different depending on the calling buffer.